+++ components/cards/index.ts
@@ -1,3 +1,6 @@
 export { SwipeCard } from "./SwipeCard";
 export { CardStack } from "./CardStack";
 export { EditModal } from "./EditModal";
+export { SideEditDrawer } from "./SideEditDrawer";
+export { GestureCard } from "./GestureCard";
+export { SelectionActionBar } from "./SelectionActionBar";